test: split AgentFS root overlay suite
